requirements.txt
mega.py==1.0.8
pycryptodome==3.15.0
pypiwin32==223